Refine types

+++ src/types/color.cr
@@ -1,10 +1,10 @@
 module Blueprint
   class Color
     JSON.mapping(
-      r: Int8,
-      g: Int8,
-      b: Int8,
-      a: Int8,
+      r: Float64,
+      g: Float64,
+      b: Float64,
+      a: Float64,
     )
 
     def initialize(@r, @g, @b, @a)
